Warning Fixed : Animation.cpp : Fixed many warnings
svn-id: r1121

+++ gemrb/plugins/Core/Animation.cpp
@@ -15,7 +15,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
  *
- * $Header: /data/gemrb/cvs2svn/gemrb/gemrb/gemrb/plugins/Core/Animation.cpp,v 1.8 2003/12/12 23:14:30 balrog994 Exp $
+ * $Header: /data/gemrb/cvs2svn/gemrb/gemrb/gemrb/plugins/Core/Animation.cpp,v 1.9 2003/12/15 09:16:47 balrog994 Exp $
  *
  */
 
@@ -108,7 +108,7 @@ Sprite2D * Animation::NextFrame(void)
 	if(pos >= indices.size())
 	pos = 0;*/
 #endif
-	if((time - starttime) >= (1000/fps)) {
+	if((time - starttime) >= (unsigned long)(1000/fps)) {
 		pos++;
 		starttime = time;
 	}
